cmd/kubenest/node-agent/requirements.txt
websockets>=9.1
pyOpenSSL>=23.2.0